feat(): implement saved search and custom email template

+++ app/src/Controller/Front/PropertyController.php
@@ -7,9 +7,18 @@ use App\Repository\SearchRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
+use Symfony\Component\Mailer\MailerInterface;
+use Symfony\Component\Mime\Email;
 use Symfony\Component\Routing\Annotation\Route;
 use App\Entity\Property;
+use App\Form\OwnedPropertiesSearchType;
 use App\Form\PropertyType;
+use App\Repository\UserRepository;
+use Symfony\Bridge\Twig\Mime\TemplatedEmail;
+use App\Security\Voter\OwnedPropertyVoter;
+use Doctrine\ORM\EntityManagerInterface;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 /**
@@ -20,24 +29,58 @@ class PropertyController extends AbstractController
     /**
      * @Route("/", name="property_index", methods={"GET"})
      */
-    public function index(PropertyRepository $propertyRepository, TokenStorageInterface $tokenStorage): Response
+    public function index(Request $request, PropertyRepository $propertyRepository): Response
     {
-        // Utilisateur courant
-        $user = $tokenStorage->getToken()->getUser();
+        $form = $this->createForm(OwnedPropertiesSearchType::class);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $formData = $form->getData();
+            $paginator = $propertyRepository->getPropertiesPaginationForUserByCity($this->getUser(), $request, $formData['city'], 4);
+        } else {
+            $paginator = $propertyRepository->getPropertiesPaginationForUser($this->getUser(), $request, 4);
+        }
+
+        foreach ($paginator as $property) {
+            $this->denyAccessUnlessGranted(OwnedPropertyVoter::VIEW, $property);
+        }
 
         return $this->render('front/property/index.html.twig', [
-            'properties' => $propertyRepository->getUserProperties($user)
+            'paginator' => $paginator,
+            'searchForm' => $form->createView()
         ]);
     }
+
+
+    /**
+     * @Route("/recherches-sauvegardees", name="property_saved_search", methods={"GET"})
+     * @param Request $request
+     * @param SearchRepository $searchRepository
+     * @return Response
+     */
+    public function saved_search(Request $request, SearchRepository $searchRepository): Response
+    {
+        $saved_searches = $searchRepository->findSavedSearchByUser();
+
+        return $this->render('front/property/saved-search.html.twig', [
+            'saved_searches' => $saved_searches,
+        ]);
+    }
+
 
     /**
      * @Route("/new", name="property_new", methods={"GET","POST"})
      * @param Request $request
      * @param SearchRepository $searchRepository
+     * @param MailerInterface $mailer
      * @return Response
+     * @throws TransportExceptionInterface
      */
-    public function new(Request $request, SearchRepository $searchRepository): Response
+    public function new(Request $request, SearchRepository $searchRepository, MailerInterface $mailer): Response
     {
+        $this->denyAccessUnlessGranted(OwnedPropertyVoter::CREATE, Property::class);
+
         $property = new Property();
         $property->setOwner($this->getUser());
         $form = $this->createForm(PropertyType::class, $property);
@@ -48,10 +91,23 @@ class PropertyController extends AbstractController
             $entityManager->persist($property);
             $entityManager->flush();
 
-            // TODO : lancer la recherche lorsqu'un nouveau bien est créé.
-            $searchRepository->findInterestedUsers($property);
+            $users = $searchRepository->findInterestedUsers($property);
 
-            // TODO : Faire une Queue pour envoyer les mails de façon asynchrone aux utilisateurs.
+            $email = (new TemplatedEmail())
+                ->from('noreply@sinequanone.com')
+                ->subject('Nouveau bien disponible')
+                ->htmlTemplate('emails/available-property.html.twig')
+            ;
+
+            foreach ($users as $user)
+            {
+                $email->to($user['email']);
+                $email->context([
+                  'firstname' => $user['firstname'],
+                  'lastname' => $user['lastname'],
+                ]);
+                $mailer->send($email);
+            }
 
             $this->addFlash('success', 'Votre nouveau bien à ' . $property->getCity() . ' s\'est ajouté correctement');
             return $this->redirectToRoute('front_property_index');
@@ -64,22 +120,12 @@ class PropertyController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="property_show", methods={"GET"})
-     * @param Property $property
-     * @return Response
-     */
-    public function show(Property $property): Response
-    {
-        return $this->render('front/property/show.html.twig', [
-            'property' => $property,
-        ]);
-    }
-
-    /**
      * @Route("/{id}/edit", name="property_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, Property $property): Response
     {
+        $this->denyAccessUnlessGranted(OwnedPropertyVoter::UPDATE, $property);
+
         $form = $this->createForm(PropertyType::class, $property);
         $form->handleRequest($request);
 
@@ -101,6 +147,8 @@ class PropertyController extends AbstractController
      */
     public function delete(Request $request, Property $property): Response
     {
+        $this->denyAccessUnlessGranted(OwnedPropertyVoter::DELETE, $property);
+
         if ($this->isCsrfTokenValid('delete' . $property->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($property);
